🏗 export/env split utils in various files
